+++ how-to/workspace-platform-starter/client/src/framework/shapes/module-shapes.ts
@@ -1,12 +1,12 @@
 import type OpenFin from "@openfin/core";
-import type { BrowserWindowModule, CustomPaletteSet, Page, Workspace } from "@openfin/workspace-platform";
+import type { BrowserWindowModule, Page } from "@openfin/workspace-platform";
 import type { PlatformApp } from "./app-shapes";
 import type { FavoriteClient } from "./favorite-shapes";
 import type { LifecycleEvents, LifecycleHandler } from "./lifecycle-shapes";
 import type { Logger, LoggerCreator } from "./logger-shapes";
 import type { PopupMenuEntry } from "./menu-shapes";
 import type { NotificationClient } from "./notification-shapes";
-import type { ColorSchemeMode } from "./theme-shapes";
+import type { ThemeClient } from "./theme-shapes";
 import type { VersionInfo } from "./version-shapes";
 
 /**
@@ -88,30 +88,6 @@ export interface ModuleHelpers {
 	getApp?(id: string): Promise<PlatformApp | undefined>;
 
 	/**
-	 * Get the current theme id.
-	 * @returns The current theme id.
-	 */
-	getCurrentThemeId(): Promise<string>;
-
-	/**
-	 * Get the current icon folder.
-	 * @returns the platform icon folder.
-	 */
-	getCurrentIconFolder(): Promise<string>;
-
-	/**
-	 * Get the current palette.
-	 * @returns The current palette.
-	 */
-	getCurrentPalette(): Promise<CustomPaletteSet>;
-
-	/**
-	 * Get the current color scheme.
-	 * @returns The current color scheme.
-	 */
-	getCurrentColorSchemeMode(): Promise<ColorSchemeMode>;
-
-	/**
 	 * Get the version information related to the platform you are running in. If you request the version info on
 	 * initialization or you execute early you might not receive all of the version related information as you may be
 	 * early. Subscribe to the life cycle event 'after-bootstrap' to ensure you have all the related versioning
@@ -129,6 +105,12 @@ export interface ModuleHelpers {
 	 * @returns The interop client.
 	 */
 	getInteropClient?(): Promise<OpenFin.InteropClient | undefined>;
+
+	/**
+	 * If this platform has been configured to support theming it will provide it.
+	 * @returns the theme client.
+	 */
+	getThemeClient(): Promise<ThemeClient>;
 
 	/**
 	 * If this platform has been configured to support favorites and you are able to receive favorites
@@ -174,11 +156,11 @@ export interface ModuleHelpers {
 
 	/**
 	 * Launch a workspace.
-	 * @param workspace The workspace to launch.
+	 * @param workspaceId The id of the workspace to launch.
 	 * @param logger Log output from the operation.
 	 * @returns Was the workspace opened.
 	 */
-	launchWorkspace?(workspace: Workspace, logger?: Logger): Promise<boolean>;
+	launchWorkspace?(workspaceId: string, logger?: Logger): Promise<boolean>;
 
 	/**
 	 * Subscribe to lifecycle events.